restructured fakes to avoid cyclic dependency
