+++ internal/prober/fakes/shoot/clientcreator.go
@@ -1,4 +1,4 @@
-package fakes
+package shoot
 
 import (
 	"context"
